Move Field to its own file
