+++ lib/bit-struct/bit-struct.rb
@@ -11,98 +11,9 @@
 class BitStruct < String
   VERSION = "0.15.0"
 
-  class Field
-    # Offset of field in bits.
-    attr_reader :offset
-    
-    # Length of field in bits.
-    attr_reader :length
-    alias size length
-    
-    # Name of field (used for its accessors).
-    attr_reader :name
-    
-    # Options, such as :default (varies for each field subclass).
-    # In general, options can be provided as strings or as symbols.
-    attr_reader :options
-    
-    # Display name of field (used for printing).
-    attr_reader :display_name
-    
-    # Default value.
-    attr_reader :default
-    
-    # Format for printed value of field.
-    attr_reader :format
-    
-    # Subclasses can override this to define a default for all fields of this
-    # class, not just the one currently being added to a BitStruct class, a
-    # "default default" if you will. The global default, if #default returns
-    # nil, is to fill the field with zero. Most field classes just let this
-    # default stand. The default can be overridden per-field when a BitStruct
-    # class is defined.
-    def self.default; nil; end
-
-    # Used in describe.
-    def self.class_name
-      @class_name ||= name[/\w+$/]
-    end
-    
-    # Used in describe. Can be overridden per-subclass, as in NestedField.
-    def class_name
-      self.class.class_name
-    end
-
-    # Yield the description of this field, as an array of 5 strings: byte
-    # offset, type, name, size, and description. The opts hash may have:
-    #
-    # :expand ::  if the value is true, expand complex fields
-    #
-    # (Subclass implementations may yield more than once for complex fields.)
-    #
-    def describe opts
-      bits = size
-      if bits > 32 and bits % 8 == 0
-        len_str = "%dB" % (bits/8)
-      else
-        len_str = "%db" % bits
-      end
-      
-      byte_offset = offset / 8 + (opts[:byte_offset] || 0)
-
-      yield ["@%d" % byte_offset, class_name, name, len_str, display_name]
-    end
-
-    # Options are _display_name_, _default_, and _format_ (subclasses of Field
-    # may add other options).
-    def initialize(offset, length, name, opts = {})
-      @offset, @length, @name, @options =
-        offset, length, name, opts
-      
-      @display_name = opts[:display_name] || opts["display_name"]
-      @default      = opts[:default] || opts["default"] || self.class.default
-      @format       = opts[:format] || opts["format"]
-    end
-    
-    # Inspect the value of this field in the specified _obj_.
-    def inspect_in_object(obj, opts)
-      val = obj.send(name)
-      str =
-        begin
-          val.inspect_with_options(opts)
-        rescue NoMethodError
-          val.inspect
-        end
-      (f=@format) ? (f % str) : str
-    end
-    
-    # Normally, all fields show up in inspect, but some, such as padding,
-    # should not.
-    def inspectable?; true; end
-  end
-  
+  require 'bit-struct/field'
   NULL_FIELD = Field.new(0, 0, :null, :display_name => "null field")
-  
+
   # Raised when a field is added after an instance has been created. Fields
   # cannot be added after this point.
   class ClosedClassError < StandardError; end
@@ -110,9 +21,9 @@ class BitStruct < String
   # Raised if the chosen field name is not allowed, either because another
   # field by that name exists, or because a method by that name exists.
   class FieldNameError < StandardError; end
-  
+
   @default_options = {}
-  
+
   @initial_value = nil
   @closed = nil
   @rest_field = nil
@@ -127,7 +38,7 @@ class BitStruct < String
         @note = nil
       end
     end
-    
+
     # ------------------------
     # :section: field access methods
     #
@@ -139,7 +50,7 @@ class BitStruct < String
     def fields
       @fields ||= self == BitStruct ? [] : superclass.fields.dup
     end
-    
+
     # Return the list of fields defined by this class, not inherited
     # from the superclass.
     def own_fields
@@ -150,9 +61,9 @@ class BitStruct < String
     def add_field(name, length, opts = {})
       round_byte_length ## just to make sure this has been calculated
       ## before adding anything
-      
+
       name = name.to_sym
-      
+
       if @closed
         raise ClosedClassError, "Cannot add field #{name}: " +
           "The definition of the #{self.inspect} BitStruct class is closed."
@@ -169,9 +80,9 @@ class BitStruct < String
           raise FieldNameError,"Field #{name} is already defined as a method."
         end
       end
-      
+
       field_class = opts[:field_class]
-      
+
       prev = fields[-1] || NULL_FIELD
       offset = prev.offset + prev.length
       field = field_class.new(offset, length, name, opts)
@@ -194,13 +105,13 @@ class BitStruct < String
     def parse_options(ary, default_name, default_field_class) # :nodoc:
       opts = ary.grep(Hash).first || {}
       opts = default_options.merge(opts)
-      
+
       opts[:display_name]  = ary.grep(String).first || default_name
       opts[:field_class]   = ary.grep(Class).first || default_field_class
-      
+
       opts
     end
-    
+
     # Get or set the hash of default options for the class, which apply to all
     # fields. Changes take effect immediately, so can be used alternatingly with
     # blocks of field declarations. If +h+ is provided, update the default
@@ -214,17 +125,17 @@ class BitStruct < String
       end
       @default_options
     end
-    
+
     # Length, in bits, of this object.
     def bit_length
       @bit_length ||= fields.inject(0) {|a, f| a + f.length}
     end
-    
+
     # Length, in bytes (rounded up), of this object.
     def round_byte_length
       @round_byte_length ||= (bit_length/8.0).ceil
     end
-    
+
     def closed! # :nodoc:
       @closed = true
     end
@@ -239,17 +150,17 @@ class BitStruct < String
       field
     end
   end
-  
+
   # Return the list of fields for this class.
   def fields
     self.class.fields
   end
-  
+
   # Return the rest field for this class.
   def rest_field
     self.class.rest_field
   end
-  
+
   # Return the field with the given name.
   def field_by_name name
     self.class.field_by_name name
@@ -266,7 +177,7 @@ class BitStruct < String
     # Default format for describe. Fields are byte, type, name, size,
     # and description.
     DESCRIBE_FORMAT = "%8s: %-12s %-14s[%4s] %s"
-    
+
     # Can be overridden to use a different format.
     def describe_format
       DESCRIBE_FORMAT
@@ -279,7 +190,7 @@ class BitStruct < String
       if fmt.kind_of? Hash
         opts = fmt; fmt = nil
       end
-      
+
       if block_given?
         fields.each do |field|
           field.describe(opts) do |desc|
@@ -287,7 +198,7 @@ class BitStruct < String
           end
         end
         nil
-        
+
       else
         fmt ||= describe_format
 
@@ -311,7 +222,7 @@ class BitStruct < String
         result
       end
     end
-    
+
     # Subclasses can use this to append a string (or several) to the #describe
     # output. Notes are not cumulative with inheritance. When used with no
     # arguments simply returns the note string
@@ -320,7 +231,7 @@ class BitStruct < String
       @note
     end
   end
-  
+
   # ------------------------
   # :section: initialization and conversion methods
   #
@@ -339,9 +250,9 @@ class BitStruct < String
       value.each do |k, v|
         send "#{k}=", v
       end
-    
+
     when nil
-      
+
     else
       if value.respond_to?(:read)
         value = value.read(self.class.round_byte_length)
@@ -349,16 +260,16 @@ class BitStruct < String
 
       self[0, value.length] = value
     end
-    
+
     self.class.closed!
     yield self if block_given?
   end
-  
+
   DEFAULT_TO_H_OPTS = {
     :convert_keys   => :to_sym,
     :include_rest   => true
   }
-  
+
   # Returns a hash of {name=>value,...} for each field. By default, include
   # the rest field.
   # Keys are symbols derived from field names using +to_sym+, unless
@@ -370,13 +281,13 @@ class BitStruct < String
     if opts[:include_rest] and (rest_field = self.class.rest_field)
       fields_for_to_h += [rest_field]
     end
-    
+
     fields_for_to_h.inject({}) do |h,f|
       h[f.name.send(converter)] = send(f.name)
       h
     end
   end
-  
+
   # Returns an array of values of the fields of the BitStruct. By default,
   # include the rest field.
   def to_a(include_rest = true)
@@ -384,13 +295,13 @@ class BitStruct < String
       fields.map do |f|
         send(f.name)
       end
-    
+
     if include_rest and (rest_field = self.class.rest_field)
       ary << send(rest_field.name)
     end
     ary
   end
-  
+
   ## temporary hack for 1.9
   if "a"[0].kind_of? String
     def [](*args)
@@ -409,7 +320,7 @@ class BitStruct < String
       end
     end
   end
-  
+
   class << self
     # The unique "prototype" object from which new instances are copied.
     # The fields of this instance can be modified in the class definition
@@ -423,7 +334,7 @@ class BitStruct < String
     #
     def initial_value   # :yields: the initial value
       unless @initial_value
-        iv = defined?(superclass.initial_value) ? 
+        iv = defined?(superclass.initial_value) ?
           superclass.initial_value.dup : ""
         if iv.length < round_byte_length
           iv << "\0" * (round_byte_length - iv.length)
@@ -432,7 +343,7 @@ class BitStruct < String
         @initial_value = "" # Serves as initval while the real initval is inited
         @initial_value = new(iv)
         @closed = false # only creating the first _real_ instance closes.
-        
+
         fields.each do |field|
           @initial_value.send("#{field.name}=", field.default) if field.default
         end
@@ -440,7 +351,7 @@ class BitStruct < String
       yield @initial_value if block_given?
       @initial_value
     end
-    
+
     # Take +data+ (a string or BitStruct) and parse it into instances of
     # the +classes+, returning them in an array. The classes can be given
     # as an array or a separate arguments. (For parsing a string into a _single_
@@ -450,7 +361,7 @@ class BitStruct < String
         c.new(data.slice!(0...c.round_byte_length))
       end
     end
-    
+
     # Join the given structs (array or multiple args) as a string.
     # Actually, the inherited String#+ instance method is the same, as is using
     # Array#join.
@@ -474,7 +385,7 @@ class BitStruct < String
     :include_class    => true,
     :simple_format    => "<%s>"
   }
-  
+
   DETAILED_INSPECT_OPTS = {
     :format           => "%s:\n%s",
     :field_format     => "%30s = %s",
@@ -485,34 +396,34 @@ class BitStruct < String
     :include_class    => true,
     :simple_format    => "\n%s"
   }
-  
+
   # A standard inspect method which does not add newlines.
   def inspect_with_options(opts = DEFAULT_INSPECT_OPTS)
     field_format = opts[:field_format]
     field_name_meth = opts[:field_name_meth]
-    
+
     fields_for_inspect = fields.select {|field| field.inspectable?}
     if opts[:include_rest] and (rest_field = self.class.rest_field)
       fields_for_inspect << rest_field
     end
-    
+
     ary = fields_for_inspect.map do |field|
       field_format %
        [field.send(field_name_meth),
         field.inspect_in_object(self, opts)]
     end
-        
+
     body = ary.join(opts[:separator])
-    
+
     if opts[:include_class]
       opts[:format] % [self.class, body]
     else
       opts[:simple_format] % body
     end
   end
-  
+
   alias inspect inspect_with_options
-  
+
   # A more visually appealing inspect method that puts each field/value on
   # a separate line. Very useful when output is scrolling by on a screen.
   #
@@ -526,7 +437,7 @@ class BitStruct < String
   # :section: field declaration methods
   #
   # ------------------------
-  
+
   # Define accessors for a variable length substring from the end of
   # the defined fields to the end of the BitStruct. The _rest_ may behave as
   # a String or as some other String or BitStruct subclass.
@@ -544,13 +455,13 @@ class BitStruct < String
   # If a hash is provided, use it for options.
   #
   # *Warning*: the rest reader method returns a copy of the field, so
-  # accessors on that returned value do not affect the original rest field. 
+  # accessors on that returned value do not affect the original rest field.
   #
   def self.rest(name, *ary)
     if @rest_field
       raise ArgumentError, "Duplicate rest field: #{name.inspect}."
     end
-    
+
     opts = parse_options(ary, name, String)
     offset = round_byte_length
     byte_range = offset..-1
@@ -563,14 +474,14 @@ class BitStruct < String
       define_method "#{name}=" do |val|
         self[byte_range] = val
       end
-      
+
       @rest_field = Field.new(offset, -1, name, {
         :display_name => opts[:display_name],
         :rest_class => field_class
       })
     end
   end
-  
+
   # Not included with the other fields, but accessible separately.
   def self.rest_field; @rest_field; end
 end
